+++ src/app/(main)/settings/about/index.tsx
@@ -1,4 +1,4 @@
-import { Logo, Tag } from '@lobehub/ui';
+import { Logo, Tag } from '@bentwnghk/ui';
 import { Flexbox } from 'react-layout-kit';
 
 import Follow from '@//features/Follow';
@@ -9,14 +9,14 @@ import { CURRENT_VERSION } from '@/const/version';
 
 const enabledTelemetryChat = getServerConfig().ENABLE_LANGFUSE;
 
-const COPYRIGHT = `© 2023-${new Date().getFullYear()} LobeHub, LLC`;
+const COPYRIGHT = `© 2023-${new Date().getFullYear()} Mr.🆖`;
 
 const Page = ({ mobile }: { mobile?: boolean }) => {
   return (
     <Flexbox align={'center'} gap={12} paddingBlock={36} width={'100%'}>
       <Logo size={mobile ? 100 : 120} />
       <h1 style={{ fontSize: mobile ? 32 : 36, fontWeight: 900, lineHeight: 1, marginBottom: 0 }}>
-        LobeChat
+      Mr.🆖 Chat
       </h1>
       <Tag>v{CURRENT_VERSION}</Tag>
       <Flexbox gap={24} style={{ marginBlock: 48 }} width={'100%'}>
@@ -24,7 +24,7 @@ const Page = ({ mobile }: { mobile?: boolean }) => {
         {enabledTelemetryChat && <Analytics />}
       </Flexbox>
       <Follow />
-      <div>Empowering your AI dreams by LobeHub</div>
+      <div>Empowering your self-directed learning with AI</div>
       <div style={{ fontWeight: 400, opacity: 0.33 }}>{COPYRIGHT}</div>
     </Flexbox>
   );