+++ CHANGELOG.md
@@ -9188,7 +9188,7 @@
 
 #### 💄 Styles
 
-- **misc**: Switch Modal components to @lobehub/ui.
+- **misc**: Switch Modal components to @bentwnghk/ui.
 
 <br />
 
@@ -9197,7 +9197,7 @@
 
 #### Styles
 
-- **misc**: Switch Modal components to @lobehub/ui ([d056015](https://github.com/lobehub/lobe-chat/commit/d056015))
+- **misc**: Switch Modal components to @bentwnghk/ui ([d056015](https://github.com/lobehub/lobe-chat/commit/d056015))
 
 </details>